requirements.txt
Pillow
pycryptodome
pyzbar
